% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thatsnotmy.R
\name{thatsnotmy}
\alias{thatsnotmy}
\title{thatsnotmy}
\usage{
thatsnotmy(silent = FALSE)
}
\arguments{
\item{silent}{Should the function print to the console?}
}
\value{
string
}
\description{
thatsnotmy
}
\examples{
thatsnotmy()
}
