% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thatsnotmy.R
\name{thatsnotmy}
\alias{thatsnotmy}
\title{thatsnotmy}
\usage{
thatsnotmy(silent = FALSE, break_char = "\\n")
}
\arguments{
\item{silent}{Should the function print to the console?}

\item{break_char}{A character to break the sentence, default is linebreak}
}
\value{
string
}
\description{
thatsnotmy
}
\examples{
thatsnotmy()
}
